Added linked list print and append methods.
