+++ src/main/java/datastructure/linkedlist/LinkedList.java
@@ -1,0 +1,48 @@
+package main.java.datastructure.linkedlist;
+
+public class LinkedList {
+    private Node head;
+    private Node tail;
+    private int length;
+    public LinkedList(int value) {
+        Node newNode = new Node(value);
+        this.head = newNode;
+        this.tail = newNode;
+        this.length = 1;
+    }
+
+    class Node {
+        int value;
+        Node next;
+        Node(int value) {
+            this.value = value;
+        }
+    }
+
+    public void print() {
+        Node temp = head;
+        while(temp != null) {
+            System.out.print(temp.value + " ");
+            temp = temp.next;
+        }
+    }
+
+    public void getHead() {
+        System.out.println("Head is : " + head.value);
+    }
+
+    public void getTail() {
+        System.out.println("Tail is : " + tail.value);
+    }
+
+    public void getLength() {
+        System.out.println("Length is : " + length);
+    }
+
+    public void append(int value) {
+        Node newNode = new Node(value);
+        tail.next = newNode;
+        tail = newNode;
+    }
+
+}
