+++ src/main/java/datastructure/linkedlist/Main.java
@@ -1,0 +1,13 @@
+package main.java.datastructure.linkedlist;
+public class Main {
+    public static void main(String... args) {
+        LinkedList linkedList = new LinkedList(4);
+        linkedList.append(5);
+        linkedList.append(6);
+
+        linkedList.getHead();
+        linkedList.getTail();
+        linkedList.getLength();
+        linkedList.print();
+    }
+}
