--- src/test/java/datastructure/linkedlist/LinkedList.java
@@ -1,21 +1,0 @@
-package test.java.datastructure.linkedlist;
-
-public class LinkedList {
-    private Node head;
-    private Node tail;
-    private int length;
-    public LinkedList(int value) {
-        Node newNode = new Node(value);
-        this.head = newNode;
-        this.tail = newNode;
-        this.length = 1;
-    }
-
-    class Node {
-        int value;
-        Node next;
-        Node(int value) {
-            this.value = value;
-        }
-    }
-}
