--- src/test/java/datastructure/linkedlist/Main.java
@@ -1,7 +1,0 @@
-package test.java.datastructure.linkedlist;
-
-public class Main {
-    public static void main(String... args) {
-        LinkedList linkedList = new LinkedList(4);
-    }
-}
